+++ src/applications/vaos/tests/e2e/page-objects/ClinicChoicePageObject.js
@@ -13,6 +13,16 @@ export class ClinicChoicePageObject extends PageObject {
     return this;
   }
 
+  /**
+   * Method to select a clinic.
+   *
+   * @param {Object} params
+   * @param {string|RegExp} [params.selection] Radio button to select
+   * @param {boolean} [params.isCovid=false]
+   *
+   * @returns
+   * @memberof ClinicChoicePageObject
+   */
   selectClinic({ selection, isCovid = false } = {}) {
     if (isCovid) {
       cy.findByText(/Choose a clinic/i, { selector: 'h1' });
