+++ src/applications/vaos/tests/e2e/workflows/appointment-list-workflow/upcoming-appointments.cypress.spec.js
@@ -1,10 +1,10 @@
 // @ts-check
 import { addDays, addMinutes, addMonths, format, subMinutes } from 'date-fns';
 import { APPOINTMENT_STATUS, VIDEO_TYPES } from '../../../../utils/constants';
-import MockAppointmentResponse from '../../fixtures/MockAppointmentResponse';
-import MockClinicResponse from '../../fixtures/MockClinicResponse';
-import MockFacilityResponse from '../../fixtures/MockFacilityResponse';
-import MockUser from '../../fixtures/MockUser';
+import MockAppointmentResponse from '../../../fixtures/MockAppointmentResponse';
+import MockClinicResponse from '../../../fixtures/MockClinicResponse';
+import MockFacilityResponse from '../../../fixtures/MockFacilityResponse';
+import MockUser from '../../../fixtures/MockUser';
 import AppointmentDetailPageObject from '../../page-objects/AppointmentList/AppointmentDetailPageObject';
 import AppointmentListPageObject from '../../page-objects/AppointmentList/AppointmentListPageObject';
 import {
