+++ src/applications/vaos/tests/e2e/workflows/community-care-workflow/podiatry.cypress.spec.js
@@ -1,10 +1,10 @@
 // @ts-check
 import { getTypeOfCareById } from '../../../../utils/appointment';
 import { PODIATRY_ID } from '../../../../utils/constants';
-import MockAppointmentResponse from '../../fixtures/MockAppointmentResponse';
-import MockFacilityResponse from '../../fixtures/MockFacilityResponse';
-import MockProviderResponse from '../../fixtures/MockProviderResponse';
-import MockUser from '../../fixtures/MockUser';
+import MockAppointmentResponse from '../../../fixtures/MockAppointmentResponse';
+import MockFacilityResponse from '../../../fixtures/MockFacilityResponse';
+import MockProviderResponse from '../../../fixtures/MockProviderResponse';
+import MockUser from '../../../fixtures/MockUser';
 import AppointmentListPageObject from '../../page-objects/AppointmentList/AppointmentListPageObject';
 import ClosestCityStatePageObject from '../../page-objects/ClosestCityStatePageObject';
 import CommunityCarePreferencesPageObject from '../../page-objects/CommunityCarePreferencesPageObject';
