+++ src/applications/vaos/tests/e2e/workflows/community-care-workflow/primary-care.cypress.spec.js
@@ -1,33 +1,33 @@
 // @ts-check
-import MockAppointmentResponse from '../../fixtures/MockAppointmentResponse';
+import { getTypeOfCareById } from '../../../../utils/appointment';
+import { APPOINTMENT_STATUS, PRIMARY_CARE } from '../../../../utils/constants';
+import MockAppointmentResponse from '../../../fixtures/MockAppointmentResponse';
+import MockFacilityResponse from '../../../fixtures/MockFacilityResponse';
+import MockProviderResponse from '../../../fixtures/MockProviderResponse';
+import MockUser from '../../../fixtures/MockUser';
+import AppointmentListPageObject from '../../page-objects/AppointmentList/AppointmentListPageObject';
+import ClosestCityStatePageObject from '../../page-objects/ClosestCityStatePageObject';
+import CommunityCarePreferencesPageObject from '../../page-objects/CommunityCarePreferencesPageObject';
+import ConfirmationPageObject from '../../page-objects/ConfirmationPageObject';
+import ContactInfoPageObject from '../../page-objects/ContactInfoPageObject';
+import DateTimeRequestPageObject from '../../page-objects/DateTimeRequestPageObject';
+import PreferredLanguagePageObject from '../../page-objects/PreferredLanguagePageObject';
+import ReasonForAppointmentPageObject from '../../page-objects/ReasonForAppointmentPageObject';
+import ReviewPageObject from '../../page-objects/ReviewPageObject';
+import TypeOfCarePageObject from '../../page-objects/TypeOfCarePageObject';
+import TypeOfFacilityPageObject from '../../page-objects/TypeOfFacilityPageObject';
 import {
-  mockAppointmentGetApi,
   mockAppointmentCreateApi,
+  mockAppointmentGetApi,
   mockAppointmentsGetApi,
   mockCCProvidersApi,
+  mockEligibilityCCApi,
   mockFacilitiesApi,
   mockFeatureToggles,
-  mockEligibilityCCApi,
   mockSchedulingConfigurationApi,
   mockVamcEhrApi,
   vaosSetup,
 } from '../../vaos-cypress-helpers';
-import MockUser from '../../fixtures/MockUser';
-import AppointmentListPageObject from '../../page-objects/AppointmentList/AppointmentListPageObject';
-import TypeOfCarePageObject from '../../page-objects/TypeOfCarePageObject';
-import TypeOfFacilityPageObject from '../../page-objects/TypeOfFacilityPageObject';
-import DateTimeRequestPageObject from '../../page-objects/DateTimeRequestPageObject';
-import CommunityCarePreferencesPageObject from '../../page-objects/CommunityCarePreferencesPageObject';
-import ReasonForAppointmentPageObject from '../../page-objects/ReasonForAppointmentPageObject';
-import ContactInfoPageObject from '../../page-objects/ContactInfoPageObject';
-import ReviewPageObject from '../../page-objects/ReviewPageObject';
-import ConfirmationPageObject from '../../page-objects/ConfirmationPageObject';
-import PreferredLanguagePageObject from '../../page-objects/PreferredLanguagePageObject';
-import { APPOINTMENT_STATUS, PRIMARY_CARE } from '../../../../utils/constants';
-import ClosestCityStatePageObject from '../../page-objects/ClosestCityStatePageObject';
-import MockProviderResponse from '../../fixtures/MockProviderResponse';
-import MockFacilityResponse from '../../fixtures/MockFacilityResponse';
-import { getTypeOfCareById } from '../../../../utils/appointment';
 
 const { cceType } = getTypeOfCareById(PRIMARY_CARE);
 const typeOfCareId = getTypeOfCareById(PRIMARY_CARE).idV2;
