+++ src/applications/vaos/tests/e2e/workflows/request-schedule-workflow/primary-care.cypress.spec.js
@@ -1,32 +1,32 @@
 // @ts-check
+import { getTypeOfCareById } from '../../../../utils/appointment';
 import { APPOINTMENT_STATUS, PRIMARY_CARE } from '../../../../utils/constants';
-import MockAppointmentResponse from '../../fixtures/MockAppointmentResponse';
+import MockAppointmentResponse from '../../../fixtures/MockAppointmentResponse';
+import MockEligibilityResponse from '../../../fixtures/MockEligibilityResponse';
+import MockFacilityResponse from '../../../fixtures/MockFacilityResponse';
+import MockUser from '../../../fixtures/MockUser';
+import AppointmentListPageObject from '../../page-objects/AppointmentList/AppointmentListPageObject';
+import ConfirmationPageObject from '../../page-objects/ConfirmationPageObject';
+import ContactInfoPageObject from '../../page-objects/ContactInfoPageObject';
+import DateTimeRequestPageObject from '../../page-objects/DateTimeRequestPageObject';
+import ReasonForAppointmentPageObject from '../../page-objects/ReasonForAppointmentPageObject';
+import ReviewPageObject from '../../page-objects/ReviewPageObject';
+import TypeOfCarePageObject from '../../page-objects/TypeOfCarePageObject';
+import TypeOfFacilityPageObject from '../../page-objects/TypeOfFacilityPageObject';
+import TypeOfVisitPageObject from '../../page-objects/TypeOfVisitPageObject';
+import VAFacilityPageObject from '../../page-objects/VAFacilityPageObject';
 import {
-  mockAppointmentGetApi,
   mockAppointmentCreateApi,
+  mockAppointmentGetApi,
   mockAppointmentsGetApi,
   mockEligibilityApi,
+  mockEligibilityCCApi,
   mockFacilitiesApi,
   mockFeatureToggles,
   mockSchedulingConfigurationApi,
   mockVamcEhrApi,
   vaosSetup,
-  mockEligibilityCCApi,
 } from '../../vaos-cypress-helpers';
-import MockUser from '../../fixtures/MockUser';
-import AppointmentListPageObject from '../../page-objects/AppointmentList/AppointmentListPageObject';
-import TypeOfCarePageObject from '../../page-objects/TypeOfCarePageObject';
-import VAFacilityPageObject from '../../page-objects/VAFacilityPageObject';
-import MockEligibilityResponse from '../../fixtures/MockEligibilityResponse';
-import MockFacilityResponse from '../../fixtures/MockFacilityResponse';
-import DateTimeRequestPageObject from '../../page-objects/DateTimeRequestPageObject';
-import TypeOfVisitPageObject from '../../page-objects/TypeOfVisitPageObject';
-import ContactInfoPageObject from '../../page-objects/ContactInfoPageObject';
-import ReviewPageObject from '../../page-objects/ReviewPageObject';
-import ConfirmationPageObject from '../../page-objects/ConfirmationPageObject';
-import ReasonForAppointmentPageObject from '../../page-objects/ReasonForAppointmentPageObject';
-import TypeOfFacilityPageObject from '../../page-objects/TypeOfFacilityPageObject';
-import { getTypeOfCareById } from '../../../../utils/appointment';
 
 const typeOfCareId = getTypeOfCareById(PRIMARY_CARE).idV2;
 const { cceType } = getTypeOfCareById(PRIMARY_CARE);
@@ -101,8 +101,10 @@ describe('VAOS request schedule flow - Primary care', () => {
           ReasonForAppointmentPageObject.assertUrl()
             .assertHeading({ name: /What.s the reason for this appointment/i })
             .selectReasonForAppointment()
-            .typeAdditionalText({
+            .assertLabel({
               label: /Add any details you.d like to share with your provider/,
+            })
+            .typeAdditionalText({
               content: 'This is a test',
             })
             .clickNextButton();
@@ -158,8 +160,10 @@ describe('VAOS request schedule flow - Primary care', () => {
           ReasonForAppointmentPageObject.assertUrl()
             .assertHeading({ name: /What.s the reason for this appointment/i })
             .selectReasonForAppointment()
-            .typeAdditionalText({
+            .assertLabel({
               label: /Add any details you.d like to share with your provider/,
+            })
+            .typeAdditionalText({
               content: 'This is a test',
             })
             .clickNextButton();
@@ -237,8 +241,10 @@ describe('VAOS request schedule flow - Primary care', () => {
           ReasonForAppointmentPageObject.assertUrl()
             .assertHeading({ name: /What.s the reason for this appointment/i })
             .selectReasonForAppointment()
-            .typeAdditionalText({
+            .assertLabel({
               label: /Add any details you.d like to share with your provider/,
+            })
+            .typeAdditionalText({
               content: 'This is a test',
             })
             .clickNextButton();
@@ -294,8 +300,10 @@ describe('VAOS request schedule flow - Primary care', () => {
           ReasonForAppointmentPageObject.assertUrl()
             .assertHeading({ name: /What.s the reason for this appointment/i })
             .selectReasonForAppointment()
-            .typeAdditionalText({
+            .assertLabel({
               label: /Add any details you.d like to share with your provider/,
+            })
+            .typeAdditionalText({
               content: 'This is a test',
             })
             .clickNextButton();
@@ -378,8 +386,10 @@ describe('VAOS request schedule flow - Primary care', () => {
         ReasonForAppointmentPageObject.assertUrl()
           .assertHeading({ name: /What.s the reason for this appointment/i })
           .selectReasonForAppointment()
-          .typeAdditionalText({
+          .assertLabel({
             label: /Add any details you.d like to share with your provider/,
+          })
+          .typeAdditionalText({
             content: 'This is a test',
           })
           .clickNextButton();
